Move files into /src
