+++ package.json
@@ -172,6 +172,9 @@
       "packages/fields",
       "packages/fields-authed-relationship",
       "packages/fields-auto-increment",
+      "packages/fields-oembed",
+      "packages/fields-unsplash",
+      "packages/fields-cloudinary-image",
       "packages/fields-datetime-utc",
       "packages/fields-markdown",
       "packages/fields-mongoid",
