+++ packages/fields-cloudinary-image/src/views/Cell.js
