+++ packages/fields-cloudinary-image/src/views/blocks/single-image.js
