+++ packages/fields-oembed/src/OEmbedBlock.js
@@ -3,7 +3,7 @@ import { importView } from '@keystonejs/build-field-types';
 
 import { Block } from '@keystonejs/field-content/Block';
 import OEmbed from '.';
-import RelationshipType from '../fields/src/types/Relationship';
+import { Relationship as RelationshipType } from '@keystonejs/fields';
 
 const RelationshipWrapper = {
   ...RelationshipType,
